Update instagram.js
This script was no longer working so I decided to fix it. 
The approach it was using (trying to select and operate only on the current <video> element being looked at) was not working, as it had the problem that it kept referencing the already gone-off-screen video element, while having no effects at all on the current "focused" video element.
My approach was to select and operate on all video elements at once. Not so elegant, but it works.